+++ src/discljord/messaging/impl.clj
@@ -820,11 +820,12 @@
   [webhook-id webhook-token message-id] :delete
   (webhook-url webhook-id webhook-token message-id))
 
-(defn- command-params [name description options default-perm]
+(defn- command-params [name description options default-perm type]
   {:body (json/write-str (cond-> {:name name
                                   :description description}
                                  options (assoc :options options)
-                                 (some? default-perm) (assoc :default_permission default-perm)))})
+                                 (some? default-perm) (assoc :default_permission default-perm)
+                                 type (assoc :type type)))})
 
 (defn- global-cmd-url
   ([application-id] (str "/applications/" application-id "/commands"))
@@ -839,16 +840,16 @@
 
 
 (defdispatch :create-global-application-command
-  [_ application-id name description] [options default-permission] _ :post status body
+  [_ application-id name description] [options default-permission type] _ :post status body
   (global-cmd-url application-id)
-  (command-params name description options default-permission)
+  (command-params name description options default-permission type)
   (cond->> (json-body body)
     (not= 2 (quot status 100)) (ex-info "Attempted to create an invalid global command")))
 
 (defdispatch :edit-global-application-command
-  [_ application-id command-id name description] [options default-permission] _ :patch status body
+  [_ application-id command-id name description] [options default-permission type] _ :patch status body
   (global-cmd-url application-id command-id)
-  (command-params name description options default-permission)
+  (command-params name description options default-permission type)
   (cond->> (json-body body)
     (not= 2 (quot status 100)) (ex-info "Attempted to edit an invalid global command")))
 
@@ -876,16 +877,16 @@
   (json-body body))
 
 (defdispatch :create-guild-application-command
-  [_ application-id guild-id name description] [options default-permission] _ :post status body
+  [_ application-id guild-id name description] [options default-permission type] _ :post status body
   (guild-cmd-url application-id guild-id)
-  (command-params name description options default-permission)
+  (command-params name description options default-permission type)
   (cond->> (json-body body)
     (not= 2 (quot status 100)) (ex-info "Attempted to create an invalid guild command")))
 
 (defdispatch :edit-guild-application-command
-  [_ application-id guild-id command-id name description] [options default-permission] _ :patch status body
+  [_ application-id guild-id command-id name description] [options default-permission type] _ :patch status body
   (guild-cmd-url application-id guild-id command-id)
-  (command-params name description options default-permission)
+  (command-params name description options default-permission type)
   (cond->> (json-body body)
     (not= 2 (quot status 100)) (ex-info "Attempted to edit an invalid guild command")))